requirements.txt

vllm  # For GPU, see: https://github.com/vllm-project/vllm#installation
torch>=2.1.2
torchvision>=0.16.2
torchaudio>=2.1.2
transformers>=4.40.0
datasets
peft
pillow==10.2.0
gradio==6.6.0
accelerate
qwen-vl-utils==0.0.8
requests>=2.31.0
spaces
numpy
spaces

numpy
